+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [**Semantic Versioning v2.0.0**](https://semver.org/
 
 ## Unreleased ##
 
+### Fixed ###
+
+* Fixed objects with getters & setters
+
 ### Security ###
 
 * Update dependencies